upload scheduler / hswitch / app1

+++ include/MCAL/gpio.h
@@ -88,7 +88,7 @@ Error_enumStatus_t GPIO_InitPin(GPIOPIN_t* gpiopin);
 
 Error_enumStatus_t GPIO_SetPinValue(void* PORT , uint32_t gpio_pinNumber , uint32_t gpiopinStatus);
 
-Error_enumStatus_t GPIO_GetPinValue(void* PORT , uint32_t gpio_pinNumber , uint32_t* gpiopinStatus);
+Error_enumStatus_t GPIO_GetPinValue(void* PORT , uint32_t gpio_pinNumber , uint8_t* gpiopinStatus);
 
 
 #endif
